+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/api/AppInfoApiSpec.kt
@@ -14,7 +14,7 @@ import retrofit2.converter.gson.GsonConverterFactory
 
 open class AppInfoApiSpec private constructor(
     internal val mockServer: MockWebServer
-) : MockWebServerBaseTest(mockServer) {
+) : MockWebServerBaseSpec(mockServer) {
 
     constructor() : this(MockWebServer())