+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/MiniAppWebviewSpec.kt
@@ -21,6 +21,7 @@ import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import com.rakuten.tech.mobile.miniapp.navigator.ExternalResultHandler
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppExternalUrlLoader
 import com.rakuten.tech.mobile.miniapp.navigator.MiniAppNavigator
+import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionCache
 import org.amshove.kluent.*
 import org.junit.Before
 import org.junit.Test
@@ -29,13 +30,14 @@ import org.mockito.Mockito
 import java.io.ByteArrayInputStream
 import kotlin.test.assertTrue
 
-open class BaseWebViewTest {
+open class BaseWebViewSpec {
     lateinit var context: Context
     lateinit var basePath: String
     internal lateinit var miniAppWebView: MiniAppWebView
     lateinit var webResourceRequest: WebResourceRequest
     val miniAppMessageBridge: MiniAppMessageBridge = mock()
     val miniAppNavigator: MiniAppNavigator = mock()
+    internal val miniAppCustomPermissionCache: MiniAppCustomPermissionCache = mock()
     internal lateinit var webChromeClient: MiniAppWebChromeClient
 
     @Before
@@ -51,14 +53,15 @@ open class BaseWebViewTest {
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = miniAppNavigator,
             hostAppUserAgentInfo = TEST_HA_NAME,
-            miniAppWebChromeClient = webChromeClient
+            miniAppWebChromeClient = webChromeClient,
+            miniAppCustomPermissionCache = miniAppCustomPermissionCache
         )
         webResourceRequest = getWebResReq(miniAppWebView.getLoadUrl().toUri())
     }
 }
 
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebviewTest : BaseWebViewTest() {
+class MiniAppWebviewSpec : BaseWebViewSpec() {
 
     @Test
     fun `for a given app id, creates corresponding view for the caller`() {
@@ -111,7 +114,8 @@ class MiniAppWebviewTest : BaseWebViewTest() {
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = miniAppNavigator,
             hostAppUserAgentInfo = "",
-            miniAppWebChromeClient = webChromeClient
+            miniAppWebChromeClient = webChromeClient,
+            miniAppCustomPermissionCache = mock()
         )
         miniAppWebView.settings.userAgentString shouldNotEndWith TEST_HA_NAME
     }
@@ -129,16 +133,24 @@ class MiniAppWebviewTest : BaseWebViewTest() {
     @Test
     fun `each mini app should have different domain`() {
         val miniAppWebViewForMiniapp1 = MiniAppWebView(
-            context, miniAppWebView.basePath, TEST_MA, miniAppMessageBridge, miniAppNavigator, TEST_HA_NAME)
+            context,
+            miniAppWebView.basePath,
+            TEST_MA,
+            miniAppMessageBridge,
+            miniAppNavigator,
+            TEST_HA_NAME,
+            mock(),
+            mock()
+        )
         val miniAppWebViewForMiniapp2 = MiniAppWebView(
             context, miniAppWebView.basePath, TEST_MA.copy(id = "app-id-2"), miniAppMessageBridge,
-            miniAppNavigator, TEST_HA_NAME)
+            miniAppNavigator, TEST_HA_NAME, mock(), mock())
         miniAppWebViewForMiniapp1.url shouldNotBeEqualTo miniAppWebViewForMiniapp2.url
     }
 
     @Test
     fun `MiniAppMessageBridge should be connected with RealMiniAppDisplay`() {
-        verify(miniAppMessageBridge, atLeastOnce()).setWebViewListener(miniAppWebView)
+        verify(miniAppMessageBridge, atLeastOnce()).init(miniAppWebView, miniAppCustomPermissionCache, TEST_MA)
     }
 
     @Test
@@ -158,7 +170,7 @@ class MiniAppWebviewTest : BaseWebViewTest() {
 
 @Suppress("SwallowedException")
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebClientTest : BaseWebViewTest() {
+class MiniAppWebClientSpec : BaseWebViewSpec() {
     private val externalResultHandler: ExternalResultHandler = spy()
     private val miniAppScheme = MiniAppScheme(TEST_MA_ID)
 
@@ -277,7 +289,7 @@ class MiniAppWebClientTest : BaseWebViewTest() {
 }
 
 @RunWith(AndroidJUnit4::class)
-class MiniAppWebChromeTest : BaseWebViewTest() {
+class MiniAppWebChromeTest : BaseWebViewSpec() {
 
     @Test
     fun `for a WebChromeClient, it should be MiniAppWebChromeClient`() {