+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/RealMiniAppDisplaySpec.kt
@@ -22,7 +22,7 @@ import org.mockito.Mockito
 
 @ExperimentalCoroutinesApi
 @RunWith(AndroidJUnit4::class)
-class RealMiniAppDisplayTest {
+class RealMiniAppDisplaySpec {
     private lateinit var context: Context
     private lateinit var basePath: String
     private lateinit var realDisplay: RealMiniAppDisplay
@@ -38,7 +38,8 @@ class RealMiniAppDisplayTest {
             miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = mock(),
-            hostAppUserAgentInfo = TEST_HA_NAME
+            hostAppUserAgentInfo = TEST_HA_NAME,
+            miniAppCustomPermissionCache = mock()
         )
     }