+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/MiniAppMessageBridgeSpec.kt
@@ -9,9 +9,7 @@ import com.rakuten.tech.mobile.miniapp.TEST_CALLBACK_ID
 import com.rakuten.tech.mobile.miniapp.TEST_CALLBACK_VALUE
 import com.rakuten.tech.mobile.miniapp.TEST_ERROR_MSG
 import com.rakuten.tech.mobile.miniapp.display.WebViewListener
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionResult
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppPermissionType
+import com.rakuten.tech.mobile.miniapp.permission.*
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
@@ -37,11 +35,21 @@ class MiniAppMessageBridgeSpec {
             }
 
             override fun requestCustomPermissions(
-                permissions: List<Pair<MiniAppCustomPermissionType, String>>,
-                callback: (grantResult: String) -> Unit
+                permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>,
+                callback: (List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>) -> Unit
             ) {
-                val grantResult = "{\"USER_NAME\":\"DENIED\"}"
+                val grantResult = "{\"rakuten.miniapp.user.USER_NAME\":\"DENIED\"}"
                 onRequestCustomPermissionsResult(TEST_CALLBACK_ID, grantResult)
+            }
+
+            override fun shareContent(
+                content: String,
+                callback: (isSuccess: Boolean, message: String?) -> Unit
+            ) {
+                callback.invoke(true, null)
+                callback.invoke(true, SUCCESS)
+                callback.invoke(false, null)
+                callback.invoke(false, TEST_ERROR_MSG)
             }
         }
 
@@ -57,6 +65,22 @@ class MiniAppMessageBridgeSpec {
         id = TEST_CALLBACK_ID)
     private val permissionJsonStr = Gson().toJson(permissionCallbackObj)
 
+    private val customPermissionCallbackObj = CustomPermissionCallbackObj(
+        action = ActionType.REQUEST_CUSTOM_PERMISSIONS.action,
+        param = CustomPermission(
+            listOf(CustomPermissionObj(MiniAppCustomPermissionType.USER_NAME.type, ""))
+        ),
+        id = TEST_CALLBACK_ID
+    )
+    private val customPermissionJsonStr = Gson().toJson(customPermissionCallbackObj)
+
+    private val shareContentCallbackObj = CallbackObj(
+        action = ActionType.SHARE_INFO.action,
+        param = ShareInfoCallbackObj.ShareInfoParam(
+            ShareInfo("This is content")),
+        id = TEST_CALLBACK_ID)
+    private val shareContentJsonStr = Gson().toJson(shareContentCallbackObj)
+
     private fun createErrorWebViewListener(errMsg: String): WebViewListener =
         object : WebViewListener {
             override fun runSuccessCallback(callbackId: String, value: String) {
@@ -70,7 +94,11 @@ class MiniAppMessageBridgeSpec {
 
     @Before
     fun setup() {
-        miniAppBridge.setWebViewListener(mock())
+        miniAppBridge.init(
+            webViewListener = mock(),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
     }
 
     @Test
@@ -85,8 +113,11 @@ class MiniAppMessageBridgeSpec {
     fun `postValue should be called when permission is granted`() {
         val isPermissionGranted = true
         val miniAppBridge = Mockito.spy(createMiniAppMessageBridge(isPermissionGranted))
-        miniAppBridge.setWebViewListener(
-            createErrorWebViewListener("Cannot request permission: null"))
+        miniAppBridge.init(
+            webViewListener = createErrorWebViewListener("Cannot request permission: null"),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
 
         miniAppBridge.postMessage(permissionJsonStr)
 
@@ -112,20 +143,56 @@ class MiniAppMessageBridgeSpec {
     @Test
     fun `postError should be called when cannot get unique id`() {
         val errMsg = "Cannot get unique id: null"
-        miniAppBridge.setWebViewListener(createErrorWebViewListener(errMsg))
+        miniAppBridge.init(
+            webViewListener = createErrorWebViewListener("Cannot get unique id: null"),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
         miniAppBridge.postMessage(uniqueIdJsonStr)
 
         verify(miniAppBridge, times(1)).postError(TEST_CALLBACK_ID, errMsg)
     }
 
     @Test
-    fun `should work without exception when there is custom permission request`() {
-        val customPermissionCallbackObj = CustomPermissionCallbackObj(
-            action = ActionType.REQUEST_CUSTOM_PERMISSIONS.action,
-            param = CustomPermission(listOf(CustomPermissionObj(MiniAppCustomPermissionType.USER_NAME.type, ""))),
-            id = TEST_CALLBACK_ID)
-        val customPermissionJsonStr = Gson().toJson(customPermissionCallbackObj)
+    fun `postValue should be called when can request custom permission`() {
+        val isPermissionGranted = false
+        val miniAppBridge = Mockito.spy(createMiniAppMessageBridge(isPermissionGranted))
+        miniAppBridge.init(
+            webViewListener = createErrorWebViewListener("Cannot request custom permissions: null"),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
 
         miniAppBridge.postMessage(customPermissionJsonStr)
+
+        verify(miniAppBridge, times(1))
+            .postValue(customPermissionCallbackObj.id, "{\"rakuten.miniapp.user.USER_NAME\":\"DENIED\"}")
+    }
+
+    @Test
+    fun `postError should be called when cannot request custom permission`() {
+        val errMsg = "Cannot request custom permissions: null"
+        miniAppBridge.init(
+            webViewListener = createErrorWebViewListener(errMsg),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
+        miniAppBridge.postMessage(customPermissionJsonStr)
+
+        verify(miniAppBridge, times(1)).postError(TEST_CALLBACK_ID, errMsg)
+    }
+
+    @Test
+    fun `postError should be called when cannot share content`() {
+        miniAppBridge.postMessage(shareContentJsonStr)
+        val errMsg = "Cannot share content: null"
+        miniAppBridge.init(
+            webViewListener = createErrorWebViewListener(errMsg),
+            customPermissionCache = mock(),
+            miniAppInfo = mock()
+        )
+        miniAppBridge.postMessage(shareContentJsonStr)
+
+        verify(miniAppBridge, times(1)).postError(TEST_CALLBACK_ID, errMsg)
     }
 }