+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/storage/UrlToFileInfoParserSpec.kt
@@ -5,7 +5,7 @@ import com.rakuten.tech.mobile.miniapp.VALID_FILE_URL_PATH
 import org.amshove.kluent.shouldEqual
 import org.junit.Test
 
-class UrlToFileInfoParserTest {
+class UrlToFileInfoParserSpec {
 
     private var urlParser = UrlToFileInfoParser()