+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/permission/CustomPermissionPresenter.kt
@@ -6,8 +6,6 @@ import android.view.LayoutInflater
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.rakuten.tech.mobile.miniapp.MiniApp
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermission
-import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionManager
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionResult
 import com.rakuten.tech.mobile.miniapp.permission.MiniAppCustomPermissionType
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.ListCustomPermissionBinding
@@ -17,83 +15,61 @@ class CustomPermissionPresenter(private val miniapp: MiniApp) {
 
     constructor() : this(MiniApp.instance(AppSettings.instance.miniAppSettings))
 
-    fun promptForCustomPermissions(
+    fun executeCustomPermissionsCallback(
         context: Context,
         miniAppId: String,
-        permissionWithDescriptions: List<Pair<MiniAppCustomPermissionType, String>>,
-        callback: (grantResult: String) -> Unit
+        permissionsWithDescription: List<Pair<MiniAppCustomPermissionType, String>>,
+        callback: (List<Pair<MiniAppCustomPermissionType, MiniAppCustomPermissionResult>>) -> Unit
     ) {
         if (miniAppId.isEmpty())
             return
 
-        // prepare UI for adapter
-        val layoutInflater = LayoutInflater.from(context)
-        val permissionLayout = ListCustomPermissionBinding.inflate(layoutInflater, null, false)
-        permissionLayout.listCustomPermission.layoutManager = LinearLayoutManager(context)
-        val adapter = CustomPermissionAdapter()
-
-        permissionLayout.listCustomPermission.addItemDecoration(
-            DividerItemDecoration(
-                context,
-                DividerItemDecoration.VERTICAL
-            )
-        )
-
-        // prepare adapter for showing items
+        // get cached data from SDK
         val cachedList = miniapp.getCustomPermissions(miniAppId).pairValues
-        val filteredPair = permissionWithDescriptions.filter { (first) ->
+
+        // prepare data for adapter to check if there is any denied permission
+        val permissionsForAdapter = permissionsWithDescription.filter { (first) ->
             cachedList.find {
                 it.first == first && it.second == MiniAppCustomPermissionResult.DENIED
             } != null
         }
 
-        val namesForAdapter: ArrayList<MiniAppCustomPermissionType> = arrayListOf()
-        val resultsForAdapter: ArrayList<MiniAppCustomPermissionResult> = arrayListOf()
-        val descriptionForAdapter: ArrayList<String> = arrayListOf()
-
-        filteredPair.forEach {
-            namesForAdapter.add(it.first)
-            descriptionForAdapter.add(it.second)
-            resultsForAdapter.add(MiniAppCustomPermissionResult.ALLOWED)
-        }
-
-        adapter.addPermissionList(namesForAdapter, resultsForAdapter, descriptionForAdapter)
-        permissionLayout.listCustomPermission.adapter = adapter
-
-        // prepare listener for adapter
-        val permissionsToStore = MiniAppCustomPermission(miniAppId, adapter.permissionPairs)
-
-        val listener = DialogInterface.OnClickListener { _, _ ->
-            miniapp.setCustomPermissions(permissionsToStore)
-
-            // send the callback with grant results
-            sendGrantResultCallback(miniAppId, permissionWithDescriptions, callback)
-        }
-
-        val permissionDialogBuilder =
-            CustomPermissionDialog.Builder().build(context).apply {
-                setView(permissionLayout.root)
-                setListener(listener)
+        // show dialog if there is any denied permission
+        if (permissionsForAdapter.isNotEmpty()) {
+            val adapter = CustomPermissionAdapter()
+            val namesForAdapter: ArrayList<MiniAppCustomPermissionType> = arrayListOf()
+            val resultsForAdapter: ArrayList<MiniAppCustomPermissionResult> = arrayListOf()
+            val descriptionForAdapter: ArrayList<String> = arrayListOf()
+            permissionsForAdapter.forEach {
+                namesForAdapter.add(it.first)
+                descriptionForAdapter.add(it.second)
+                resultsForAdapter.add(MiniAppCustomPermissionResult.ALLOWED)
             }
 
-        // show dialog if there is any denied permission,
-        // otherwise send the callback with grant results
-        if (filteredPair.isNotEmpty())
-            permissionDialogBuilder.show()
-        else
-            sendGrantResultCallback(miniAppId, permissionWithDescriptions, callback)
+            adapter.addPermissionList(namesForAdapter, resultsForAdapter, descriptionForAdapter)
+            val permissionLayout = getPermissionLayout(context)
+            permissionLayout.listCustomPermission.adapter = adapter
+
+            // show dialog with listener which will invoke the callback
+            CustomPermissionDialog.Builder().build(context).apply {
+                setView(permissionLayout.root)
+                setListener(DialogInterface.OnClickListener { _, _ ->
+                    callback.invoke(adapter.permissionPairs)
+                })
+            }.show()
+        } else {
+            callback.invoke(cachedList)
+        }
     }
 
-    private fun sendGrantResultCallback(
-        miniAppId: String,
-        permissionWithDescriptions: List<Pair<MiniAppCustomPermissionType, String>>,
-        callback: (grantResult: String) -> Unit
-    ) {
-        // send json response to miniapp
-        val result = MiniAppCustomPermissionManager(miniapp).createJsonResponse(
-            miniAppId,
-            permissionWithDescriptions
+    private fun getPermissionLayout(context: Context): ListCustomPermissionBinding {
+        val layoutInflater = LayoutInflater.from(context)
+        val permissionLayout = ListCustomPermissionBinding.inflate(layoutInflater, null, false)
+        permissionLayout.listCustomPermission.layoutManager = LinearLayoutManager(context)
+        permissionLayout.listCustomPermission.addItemDecoration(
+            DividerItemDecoration(context, DividerItemDecoration.VERTICAL)
         )
-        callback.invoke(result)
+
+        return permissionLayout
     }
 }