+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/userdata/ContactsActivity.kt
@@ -9,35 +9,19 @@ import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.testapp.AppScreen
+import com.rakuten.tech.mobile.testapp.helper.clearWhiteSpaces
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
 import com.rakuten.tech.mobile.testapp.ui.settings.SettingsMenuActivity
 import kotlinx.android.synthetic.main.contacts_activity.*
+import java.util.UUID
 
 class ContactsActivity : BaseActivity() {
 
-    companion object {
-        fun start(activity: Activity) {
-            val intent = Intent(activity, ContactsActivity::class.java)
-            intent.putExtra(
-                SettingsMenuActivity.SETTINGS_SCREEN_NAME,
-                AppScreen.MINI_APP_SETTINGS_ACTIVITY
-            )
-            activity.startActivity(intent)
-        }
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
         setContentView(R.layout.contacts_activity)
         initializeActionBar()
-        setContactList()
-    }
-
-    private fun initializeActionBar() {
-        val toolBar = findViewById<Toolbar>(R.id.toolbar)
-        setSupportActionBar(toolBar)
-        showBackIcon()
+        createContactList()
     }
 
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
@@ -50,19 +34,25 @@ class ContactsActivity : BaseActivity() {
         }
     }
 
-    private fun setContactList() {
-        val adapter = ContactsAdapter()
-        val dummyContacts = arrayListOf<String>()
-        for (i in 1..10) dummyContacts.add("User Contact - $i")
-        adapter.addContactList(dummyContacts)
-        listContacts.adapter = adapter
+    private fun initializeActionBar() {
+        val toolBar = findViewById<Toolbar>(R.id.toolbar)
+        setSupportActionBar(toolBar)
+        showBackIcon()
+    }
 
+    private fun createContactList() {
+        // prepare random contact list
+        val randomContacts = arrayListOf<String>()
+        for (i in 1..10)
+            randomContacts.add(clearWhiteSpaces((UUID.randomUUID().toString())))
+
+        // add contacts in adapter
+        val adapter = ContactsAdapter()
+        adapter.addContactList(randomContacts)
+        listContacts.adapter = adapter
         listContacts.layoutManager = LinearLayoutManager(applicationContext)
         listContacts.addItemDecoration(
-            DividerItemDecoration(
-                applicationContext,
-                DividerItemDecoration.VERTICAL
-            )
+            DividerItemDecoration(this, DividerItemDecoration.VERTICAL)
         )
     }
 
@@ -79,8 +69,14 @@ class ContactsActivity : BaseActivity() {
         }
     }
 
-    override fun onBackPressed() {
-        navigateToPreviousScreen()
-        super.onBackPressed()
+    companion object {
+        fun start(activity: Activity) {
+            val intent = Intent(activity, ContactsActivity::class.java)
+            intent.putExtra(
+                SettingsMenuActivity.SETTINGS_SCREEN_NAME,
+                AppScreen.MINI_APP_SETTINGS_ACTIVITY
+            )
+            activity.startActivity(intent)
+        }
     }
 }