+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ### 2.x.x (in progress)
 **SDK**
 - **Feature:** Added public interface to read and store the custom permissions inside Android SDK. 
+- **Feature:** Share content from mini app. 
 
 **Sample App**
 - **Feature:** Added example for showing list of custom permissions (on request from Mini app) and response back to Mini app.
+- **Feature:** User name, profile photo, and contact list can be configured in the settings screen.
 
 ### 2.1.0 (2020-09-03)
 **SDK**